fix: remove types into src/types and pass props to app-screen
