+++ project/src/components/film-card/film-card.tsx
@@ -1,19 +1,13 @@
-type Title = string;
-type Img = string;
+import {FilmCardProps} from '../../types/types';
 
-type FilmCardProps = {
-  title: Title,
-  img: Img,
-}
-
-function FilmCard(props: FilmCardProps) {
+function FilmCard({img, title}: FilmCardProps) {
   return (
     <article className="small-film-card catalog__films-card">
       <div className="small-film-card__image">
-        <img src={props.img} alt={props.title} width="280" height="175" />
+        <img src={img} alt={title} width="280" height="175" />
       </div>
       <h3 className="small-film-card__title">
-        <a className="small-film-card__link" href="film-page.html">{props.title}</a>
+        <a className="small-film-card__link" href="film-page.html">{title}</a>
       </h3>
     </article>
   );
